requirements.txt
streamlit==1.28.1
Pillow==10.0.1
google-generativeai==0.3.0
python-dotenv==1.0.0
